Linting, added useBlockProps, and fixed some spelling and gramatical errors.
The code in edit.js and preview.js was not properly formatted, sometimes in the incorrect order, and had some spelling mistakes in the view render. I ran ES Lint on both edit and preview, added useBlockProps to the edit view so that the block was properly wrapped - which turned the toolbar back on. I also renamed the icon variables to be more descriptive to potentially cut down on any confusion in the future.