Added apiFetch, changed how postId is determined
Related posts were being fetched using the window.fetch method rather than WordP
ress' apiFetch. I added @wordpress/api-fetch as a dev dependency, rewrote the ca
ll to the REST API to use that. Also, the block was trying to save the incorrect response for the posts. The postId was being determined using the
URL and regex. I'm not certain, but this seems like it could cause an issue. So
instead I am pulling it in via the context attribute.